Remove debug code and fix DxStation amplitude calculation
- Remove all diagnostic/debug code added during audio investigation
- Fix DxStation amplitude calculation (parentheses bug: was computing
  sin(M_PI * random - 0.5), should be sin(M_PI * (random - 0.5)))
- Change audio output from double to float for proper RtAudio format
- Add clarifying comment about AGC valbuf indexing
- Improve pitch offset clamping using std::clamp instead of fmod

The amplitude bug was causing incorrect signal distribution (18k-42k
instead of 9k-45k range). Audio harshness with multiple stations
remains under investigation.

+++ contest.h
@@ -137,7 +137,7 @@ private:
 	                         unsigned int status, void *userData);
 
 	// Instance audio processing method
-	void getAudio(double *outdata, unsigned int nframes);
+	void getAudio(float *outdata, unsigned int nframes);
 
 	// RF gain function for QSK
 	double rfgfun(double a0, double a1);
